Major refactoring
Refactored: register, login and logout.
Moved Register, login and logout in their specific controller file, following the MVC pattern.
Created the verifyToken middleware that in the future will be used to authenticate the user that wants to do something in the chat. Created a chat function but did not implement it yet